ModelService/Main/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2


starlette>=0.14.2


chromadb>=0.3.0
pypdf2>=2.0.0
python-docx>=0.8.11

scikit-learn>=0.24.2
jieba>=0.42.1
numpy>=1.21.0


requests>=2.26.0


python-dotenv>=0.19.0


websockets>=10.0

django>=3.2.0
channels>=3.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.8.0 